+++ main.go
@@ -6,10 +6,7 @@ import (
 	"html/template"
 	"log"
 	"net/http"
-	"os"
-	"os/signal"
 	"sync"
-	"syscall"
 	"time"
 
 	"github.com/gorilla/websocket"
@@ -221,24 +218,4 @@ func getFeedsHandler(w http.ResponseWriter, r *http.Request) {
 
 	w.Header().Set("Content-Type", "application/json")
 	json.NewEncoder(w).Encode(feeds)
-}
-
-func handleSignal() {
-	ch := make(chan os.Signal)
-	signal.Notify(ch, syscall.SIGUSR1)
-	defer close(ch)
-
-	for {
-		<-ch
-		conf, err := parseConf()
-		if err != nil {
-			continue
-		}
-		increment := rssUrls.getIncrement(conf)
-		rssUrls = conf
-		now := time.Now().Format("2006-01-02 15:04:05")
-		for _, item := range increment {
-			go updateFeed(item, now)
-		}
-	}
 }
