+++ signal_windows.go
@@ -1,0 +1,8 @@
+//go:build windows
+// +build windows
+
+package main
+
+func handleSignal() {
+
+}
